[Tensor] Seperate TensorBase and TensorConcept
This PR seperates the definition of tensor base and tensor concept classes from tensor_v2.cpp.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped
